@@ -1,50 +1,49 @@
-import 'package:bloc/bloc.dart';
 import 'package:firebase_core/firebase_core.dart';
+import 'package:firebase_messaging/firebase_messaging.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:socialApp/Module/onBoarding/onBoarding.dart';
-import 'package:socialApp/shared/bloc_observer.dart';
-import 'package:socialApp/shared/components/constants.dart';
-import 'package:socialApp/shared/cubit/social_cubit.dart';
-import 'package:socialApp/shared/network/local/cache_helper.dart';
-import 'package:socialApp/shared/styles/themes.dart';
-
-import 'Module/login/social_login_screen.dart';
-import 'layout/social_layout.dart';
+import 'package:twasol/layout/splash/splash_screen.dart';
+import 'package:twasol/shared/bloc_observer.dart';
+import 'package:twasol/shared/components/constants.dart';
+import 'package:twasol/shared/cubit/social_cubit.dart';
+import 'package:twasol/shared/network/local/cache_helper.dart';
+import 'package:twasol/shared/styles/themes.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
   await Firebase.initializeApp();
   await CacheHelper.init();
 
-  Widget widget;
+  var r = await FirebaseMessaging.instance.getToken();
+  print('=================token=============================');
+  print(r);
+
   // bool onBoarding = false;
   uId = CacheHelper.getData(key: 'uId');
-  if (uId != null) {
-    widget = const SocialLayout();
-  } else {
-    widget = const OnBoardingScreen();
-  }
 
   BlocOverrides.runZoned(
     () {
-      runApp(MyApp(widget));
+      runApp(splashScreen());
     },
     blocObserver: MyBlocObserver(),
   );
 }
 
 class MyApp extends StatelessWidget {
-  final Widget startWidget;
+  final Widget? startWidget;
 
-  const MyApp(this.startWidget, {Key? key}) : super(key: key);
+  MyApp({
+     this.startWidget,
+  });
 
   @override
   Widget build(BuildContext context) {
     return MultiBlocProvider(
       providers: [
         BlocProvider(
-          create: (context) => SocialCubit()..getDataUser()..getPosts(),
+          create: (context) => SocialCubit()
+            ..getDataUser()
+            ..getPosts(),
         )
       ],
       child: MaterialApp(
