@@ -1,19 +1,18 @@
 import 'dart:io';
-
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:image_picker/image_picker.dart';
-import 'package:socialApp/Module/nav_bar/chats/chats_screen.dart';
-import 'package:socialApp/Module/nav_bar/home/feeds_screen.dart';
-import 'package:socialApp/Module/nav_bar/post/post_screen.dart';
-import 'package:socialApp/Module/nav_bar/setting/setting_screen.dart';
-import 'package:socialApp/Module/nav_bar/users/users_screen.dart';
-import 'package:socialApp/model/massege_model.dart';
-import 'package:socialApp/model/post_model.dart';
-import 'package:socialApp/model/user_model.dart';
-import 'package:socialApp/shared/components/constants.dart';
-import 'package:socialApp/shared/cubit/social_states.dart';
+import 'package:twasol/Module/nav_bar/chats/chats_screen.dart';
+import 'package:twasol/Module/nav_bar/home/feeds_screen.dart';
+import 'package:twasol/Module/nav_bar/post/post_screen.dart';
+import 'package:twasol/Module/nav_bar/setting/setting_screen.dart';
+import 'package:twasol/Module/nav_bar/users/users_screen.dart';
+import 'package:twasol/model/massege_model.dart';
+import 'package:twasol/model/post_model.dart';
+import 'package:twasol/model/user_model.dart';
+import 'package:twasol/shared/components/constants.dart';
+import 'package:twasol/shared/cubit/social_states.dart';
 import 'package:firebase_storage/firebase_storage.dart' as firebase_storage;
 
 class SocialCubit extends Cubit<SocialStates> {
@@ -72,7 +71,7 @@ class SocialCubit extends Cubit<SocialStates> {
 
   Future<void> getProfileImage() async {
     final pickedFile =
-    await profileImagePicker.pickImage(source: ImageSource.gallery);
+        await profileImagePicker.pickImage(source: ImageSource.gallery);
     if (pickedFile != null) {
       profileImage = File(pickedFile.path);
 
@@ -89,7 +88,7 @@ class SocialCubit extends Cubit<SocialStates> {
 
   Future<void> getCoverImage() async {
     final pickedFile =
-    await coverImagePicker.pickImage(source: ImageSource.gallery);
+        await coverImagePicker.pickImage(source: ImageSource.gallery);
     if (pickedFile != null) {
       coverImage = File(pickedFile.path);
       emit(SocialCoverImageSuccessStates());
@@ -101,11 +100,11 @@ class SocialCubit extends Cubit<SocialStates> {
 
   ImageProvider? putProfileImage() {
     if (coverImage == null) {
-      NetworkImage(
+      return NetworkImage(
         '${userModel!.image}',
       );
     } else {
-      FileImage(File(profileImage!.path));
+      return FileImage(File(profileImage!.path));
     }
   }
 
@@ -117,24 +116,20 @@ class SocialCubit extends Cubit<SocialStates> {
   }) {
     firebase_storage.FirebaseStorage.instance
         .ref()
-        .child('users/${Uri
-        .file(profileImage!.path)
-        .pathSegments
-        .last}')
+        .child('users/${Uri.file(profileImage!.path).pathSegments.last}')
         .putFile(profileImage!)
-        .then((value) =>
-    {
-      value.ref.getDownloadURL().then((value) {
-        updateUser(
-          userName: userName,
-          bio: bio,
-          image: value,
-        );
-        emit(SocialUploadProfileImageSuccessStates());
-      }).catchError((error) {
-        emit(SocialUploadProfileImageErrorStates());
-      })
-    })
+        .then((value) => {
+              value.ref.getDownloadURL().then((value) {
+                updateUser(
+                  userName: userName,
+                  bio: bio,
+                  image: value,
+                );
+                emit(SocialUploadProfileImageSuccessStates());
+              }).catchError((error) {
+                emit(SocialUploadProfileImageErrorStates());
+              })
+            })
         .catchError((error) {
       emit(SocialUploadProfileImageErrorStates());
     });
@@ -148,25 +143,21 @@ class SocialCubit extends Cubit<SocialStates> {
   }) {
     firebase_storage.FirebaseStorage.instance
         .ref()
-        .child('users/${Uri
-        .file(coverImage!.path)
-        .pathSegments
-        .last}')
+        .child('users/${Uri.file(coverImage!.path).pathSegments.last}')
         .putFile(coverImage!)
-        .then((value) =>
-    {
-      value.ref.getDownloadURL().then((value) {
-        updateUser(
-          userName: userName,
-          bio: bio,
-          cover: value,
-        );
+        .then((value) => {
+              value.ref.getDownloadURL().then((value) {
+                updateUser(
+                  userName: userName,
+                  bio: bio,
+                  cover: value,
+                );
 
-        emit(SocialUploadCoverImageSuccessStates());
-      }).catchError((error) {
-        emit(SocialUploadCoverImageErrorStates());
-      })
-    })
+                emit(SocialUploadCoverImageSuccessStates());
+              }).catchError((error) {
+                emit(SocialUploadCoverImageErrorStates());
+              })
+            })
         .catchError((error) {
       emit(SocialUploadCoverImageErrorStates());
     });
@@ -189,11 +180,7 @@ class SocialCubit extends Cubit<SocialStates> {
       uId: userModel!.uId,
     );
 
-    FirebaseFirestore.instance
-        .collection('users')
-        .doc(uId)
-        .update(model.toMap())
-        .then((value) {
+    FirebaseFirestore.instance.collection('users').doc(uId).update(model.toMap()).then((value) {
       getDataUser();
       emit(SocialUpdateSuccessfullySuccessStates());
     }).catchError((error) {
@@ -207,7 +194,7 @@ class SocialCubit extends Cubit<SocialStates> {
 
   Future<void> getPostImage() async {
     final pickedFile =
-    await postImagePicker.pickImage(source: ImageSource.gallery);
+        await postImagePicker.pickImage(source: ImageSource.gallery);
     if (pickedFile != null) {
       postImage = File(pickedFile.path);
       emit(SocialPostImagePickedSuccessStates());
@@ -225,30 +212,27 @@ class SocialCubit extends Cubit<SocialStates> {
 
     firebase_storage.FirebaseStorage.instance
         .ref()
-        .child('posts/${Uri
-        .file(postImage!.path)
-        .pathSegments
-        .last}')
+        .child('posts/${Uri.file(postImage!.path).pathSegments.last}')
         .putFile(postImage!)
-        .then((value) =>
-    {
-      value.ref.getDownloadURL().then((value) {
-        print(value);
+        .then((value) => {
+              value.ref.getDownloadURL().then((value) {
+                print(value);
 
-        createPost(
-          dateTime: dateTime,
-          text: text,
-          postImage: value,
-        );
-        emit(SocialCreatePostSuccessStates());
-      }).catchError((error) {
-        emit(SocialCreatePostErrorStates());
-      })
-    })
+                createPost(
+                  dateTime: dateTime,
+                  text: text,
+                  postImage: value,
+                );
+                emit(SocialCreatePostSuccessStates());
+              }).catchError((error) {
+                emit(SocialCreatePostErrorStates());
+              })
+            })
         .catchError((error) {
       emit(SocialCreatePostErrorStates());
     });
   }
+
 
   void createPost({
     required String dateTime,
@@ -265,10 +249,7 @@ class SocialCubit extends Cubit<SocialStates> {
         text: text,
         postImage: postImage ?? '');
 
-    FirebaseFirestore.instance
-        .collection('posts')
-        .add(model.toMap())
-        .then((value) {
+    FirebaseFirestore.instance.collection('posts').add(model.toMap()).then((value) {
       emit(SocialCreatePostSuccessStates());
     }).catchError((error) {
       emit(SocialCreatePostErrorStates());
@@ -290,9 +271,10 @@ class SocialCubit extends Cubit<SocialStates> {
 
     FirebaseFirestore.instance
         .collection('posts')
-        .orderBy('dateTime')
+        .orderBy('dateTime', descending: true)
         .get()
         .then((value) {
+      posts = [];
       emit(SocialLikePostsLoadingStates());
 
       value.docs.forEach((element) {
@@ -322,40 +304,25 @@ class SocialCubit extends Cubit<SocialStates> {
     });
   }
 
-  // void getStreamPosts() {
-  //   emit(SocialGetPostsDataLoadingStates());
-  //
-  //   FirebaseFirestore.instance
-  //       .collection('posts')
-  //       .orderBy('dateTime')
-  //       .snapshots().listen((event) {
-  //
-  //     event.docs.forEach((element)
-  //     {
-  //       element.reference
-  //           .collection('likes')
-  //           .snapshots()
-  //           .listen((event)
-  //         {
-  //           likes.add(event.docs.length);
-  //           postsId.add(element.id);
-  //         });
-  //       posts.add(PostModel.fromJson(element.data()));
-  //
-  //
-  //
-  //       element.reference
-  //           .collection('comments')
-  //           .snapshots()
-  //           .listen((event) {
-  //         comments.add(event.docs.length);
-  //       });
-  //
-  //         emit(SocialCommentsPostsSuccessStates());
-  //       });
-  //     });
-  //       }
+  void getLivePosts() {
+    emit(SocialGetPostsDataLoadingStates());
 
+    FirebaseFirestore.instance
+        .collection('posts')
+        .orderBy(
+          'dataTime',
+          descending: true,
+        )
+        .snapshots()
+        .listen((event) {
+      posts = [];
+      event.docs.forEach((element) {
+        postsId.add(element.id);
+        posts.add(PostModel.fromJson(element.data()));
+        emit(SocialGetPostsDataSuccessStates());
+      });
+    });
+  }
 
   void likePost(String postId) {
     FirebaseFirestore.instance
@@ -388,10 +355,7 @@ class SocialCubit extends Cubit<SocialStates> {
   void getAllUsers() {
     emit(SocialGetAllUsersLoadingStates());
     if (users.isEmpty) {
-      FirebaseFirestore.instance
-          .collection('users')
-          .get()
-          .then((value) {
+      FirebaseFirestore.instance.collection('users').get().then((value) {
         value.docs.forEach((element) {
           if (element.data()['uId'] != userModel!.uId) {
             users.add(UserModel.fromJson(element.data()));
@@ -450,7 +414,6 @@ class SocialCubit extends Cubit<SocialStates> {
   }) {
     emit(SocialGetMessagesLoadingStates());
 
-
     FirebaseFirestore.instance
         .collection('users')
         .doc(userModel!.uId)
@@ -460,15 +423,23 @@ class SocialCubit extends Cubit<SocialStates> {
         .orderBy('dataTime')
         .snapshots()
         .listen((event) {
-
-        messages = [];
-        print(messages);
-        event.docs.forEach((element) {
-          messages.add(MessageModel.fromJson(element.data()));
+      messages = [];
+      print(messages);
+      event.docs.forEach((element) {
+        messages.add(MessageModel.fromJson(element.data()));
       });
 
       emit(SocialGetMessagesSuccessStates());
     });
   }
 
+  Future deletePost({
+    required PostModel post,
+  }) async{
+    await FirebaseFirestore.instance.collection('posts').doc(post.uId).delete().then((value) {
+      emit(SocialDeletePostSuccessStates());
+    }).catchError((error) {
+      emit(SocialDeletePostErrorStates(error.toString()));
+    });
+  }
 }
